코드 리뷰 적용

+++ apps/mobile/src/screens/Settings/Profile.tsx
@@ -4,7 +4,7 @@ import {
   authControllerSignOut,
   userControllerGetUserProfile,
 } from '@repo/api-client';
-import { useMutation, useQuery } from '@tanstack/react-query';
+import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 import { useAuthStore } from '../../store/auth';
 import Loading from '../../components/common/Loading';
 import Toast from '@/components/common/Toast';
@@ -12,51 +12,72 @@ import TouchableButton from '@/components/common/TouchableButton';
 import Popup from '@/components/common/Popup';
 
 const Profile = () => {
+  const queryClient = useQueryClient();
+
   const { data: userProfile } = useQuery({
     queryKey: [userControllerGetUserProfile.name],
     queryFn: userControllerGetUserProfile,
     select: data => data.data.data,
   });
 
-  const { mutate: signOut, isPending: isSignOutPending } = useMutation({
+  const { mutateAsync: signOut, isPending: isSignOutPending } = useMutation({
     mutationFn: authControllerSignOut,
-    onSuccess: () => {
-      useAuthStore.getState().clear();
-      Loading.close();
-      Toast.show('로그아웃에 성공했습니다.');
-    },
-    onError: () => {
-      Loading.close();
-      Toast.show('로그아웃에 실패했습니다. 다시 시도해주세요.');
-    },
   });
 
-  const { mutate: deleteAccount } = useMutation({
-    mutationFn: authControllerDeleteAccount,
-    onSuccess: () => {
+  const { mutateAsync: deleteAccount, isPending: isDeleteAccountPending } =
+    useMutation({
+      mutationFn: authControllerDeleteAccount,
+    });
+
+  const handleDeleteAccount = async () => {
+    if (isDeleteAccountPending) return;
+
+    try {
+      Loading.show();
+      await deleteAccount();
+
+      queryClient.removeQueries({
+        queryKey: [userControllerGetUserProfile.name],
+      });
       useAuthStore.getState().clear();
-      Loading.close();
       Toast.show('회원 탈퇴에 성공했습니다.');
-    },
-    onError: () => {
-      Loading.close();
+    } catch (error) {
       Toast.show('회원 탈퇴에 실패했습니다. 다시 시도해주세요.');
-    },
-  });
+    } finally {
+      Loading.close();
+    }
+  };
 
-  const handleDeleteAccount = () => {
+  const showDeleteAccountPopup = () => {
     Popup.show({
       title: '회원 탈퇴',
       description:
         '회원 탈퇴 하시겠습니까?\n회원 탈퇴 시 모든 정보가 삭제됩니다.',
       rightButton: {
         title: '회원 탈퇴',
-        onPress: () => {
-          Loading.show();
-          deleteAccount();
-        },
+        onPress: handleDeleteAccount,
       },
     });
+  };
+
+  const handleSignOut = async () => {
+    if (isSignOutPending) return;
+
+    try {
+      Loading.show();
+      await signOut();
+
+      useAuthStore.getState().clear();
+      queryClient.removeQueries({
+        queryKey: [userControllerGetUserProfile.name],
+      });
+
+      Toast.show('로그아웃에 성공했습니다.');
+    } catch (error) {
+      Toast.show('로그아웃에 실패했습니다. 다시 시도해주세요.');
+    } finally {
+      Loading.close();
+    }
   };
 
   if (!userProfile) return null;
@@ -69,17 +90,9 @@ const Profile = () => {
         <Text>{userProfile?.email}</Text>
       </View>
 
-      <TouchableButton
-        label="로그아웃"
-        onPress={() => {
-          if (isSignOutPending) return;
+      <TouchableButton label="로그아웃" onPress={handleSignOut} />
 
-          Loading.show();
-          signOut();
-        }}
-      />
-
-      <TouchableButton label="회원 탈퇴" onPress={handleDeleteAccount} />
+      <TouchableButton label="회원 탈퇴" onPress={showDeleteAccountPopup} />
     </View>
   );
 };
